// frontend/src/app/models/realestate.ts
export class Realestate {
    name: string;
    city: string;
    municipality: string;
    street: string;
    number: string;
    realestateType: number;
    furnished: boolean;
    floor: number;
    floors: number;
    area: number;
    rooms: number;
    media: Array<Object>;
    adType: number;
    cost: number;
    owner: string;
    pending: boolean;
    promoted: boolean;
    _id: string;
}